+++ src/main/kotlin/app/revanced/patches/youtube/layout/buttons/navigation/patch/ResolvePivotBarFingerprintsPatch.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.youtube.layout.buttons.pivotbar.shared.patch
+package app.revanced.patches.youtube.layout.buttons.navigation.patch
 
 import app.revanced.extensions.toErrorResult
 import app.revanced.patcher.annotation.Description
@@ -11,21 +11,15 @@ import app.revanced.patcher.patch.PatchResultError
 import app.revanced.patcher.patch.PatchResultSuccess
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patches.shared.mapping.misc.patch.ResourceMappingPatch
-import app.revanced.patches.youtube.layout.buttons.pivotbar.shared.annotations.PivotBarCompatibility
-import app.revanced.patches.youtube.layout.buttons.pivotbar.shared.fingerprints.InitializeButtonsFingerprint
-import app.revanced.patches.youtube.layout.buttons.pivotbar.shared.fingerprints.PivotBarConstructorFingerprint
+import app.revanced.patches.youtube.layout.buttons.navigation.fingerprints.InitializeButtonsFingerprint
+import app.revanced.patches.youtube.layout.buttons.navigation.fingerprints.PivotBarConstructorFingerprint
 
 @DependsOn([ResourceMappingPatch::class])
-@PivotBarCompatibility
 @Description("Resolves necessary fingerprints.")
 @Version("0.0.1")
 class ResolvePivotBarFingerprintsPatch : BytecodePatch(
     listOf(PivotBarConstructorFingerprint)
 ) {
-    internal companion object {
-        var imageOnlyTabResourceId: Long = -1
-    }
-
     override fun execute(context: BytecodeContext): PatchResult {
         // imageOnlyTabResourceId is used in InitializeButtonsFingerprint fingerprint
         ResourceMappingPatch.resourceMappings.find { it.type == "layout" && it.name == "image_only_tab" }
@@ -41,5 +35,9 @@ class ResolvePivotBarFingerprintsPatch : BytecodePatch(
             ) return InitializeButtonsFingerprint.toErrorResult()
         } ?: return PivotBarConstructorFingerprint.toErrorResult()
         return PatchResultSuccess()
+    }
+
+    internal companion object {
+        var imageOnlyTabResourceId: Long = -1
     }
 }