+++ src/main/kotlin/app/revanced/patches/youtube/layout/tabletminiplayer/fingerprints/MiniPlayerOverrideFingerprint.kt
@@ -7,5 +7,17 @@ import org.jf.dexlib2.Opcode
 
 object MiniPlayerOverrideFingerprint : MethodFingerprint(
     "Z", AccessFlags.STATIC or AccessFlags.PUBLIC,
-    opcodes = listOf(Opcode.RETURN), // anchor to insert the instruction
+    listOf("L"),
+    opcodes = listOf(
+        Opcode.INVOKE_STATIC,
+        Opcode.MOVE_RESULT,
+        Opcode.CONST_4,
+        Opcode.IF_EQ,
+        Opcode.CONST_4,
+        Opcode.IF_EQ,
+        Opcode.CONST_4, // override this value
+        Opcode.RETURN,
+        Opcode.CONST_4, // override this value
+        Opcode.RETURN
+    ),
 )